Refactoring (new attention module). Additional tests for FlashInfer SDPA kernel (#95)
