+++ keys_values/kvcache/base.py
@@ -23,7 +23,7 @@ from keys_values.attention import (
     KeysAndValues,
     MultiHeadSelfAttention,
 )
-from keys_values.use_eager_kernel import transform_mha_kwargs
+from keys_values.attention.use_eager_kernel import transform_mha_kwargs
 
 
 @dataclass(frozen=True)
