+++ keys_values/scripts/profile_naive_vs_padded_sdpa.py
@@ -22,14 +22,14 @@ import torch
 
 from keys_values.config import Config
 
-from keys_values.attention import (
+from keys_values.attention.attention_utils import SDPA_KERNELS_BEST_ORDERING
+from keys_values.attention.base import (
     scaled_dot_product_attention_in_blocks,
     DefaultKeysAndValues,
 )
-from keys_values.attention_utils import SDPA_KERNELS_BEST_ORDERING
 from keys_values.kvcache.base import KVCacheParams
 from keys_values.kvcache.test_utils import random_keys_values, random_tensor
-from keys_values.sdpa_wrapper import scaled_dot_product_attention
+from keys_values.attention.sdpa_wrapper import scaled_dot_product_attention
 from keys_values.utils import append_results_to_csv, randint_torch
