+++ test/attention/test_attention.py
@@ -29,20 +29,20 @@ from litgpt.model import (
 from litgpt.utils import _RunIf
 
 from keys_values.array_limit import TemporaryArrayLimit
-from keys_values.attention import (
-    DefaultKeysAndValues,
-    MultiHeadSelfAttention,
-    do_softcapping,
-    eager_scaled_dot_product_attention,
-)
-from keys_values.attention_utils import (
+from keys_values.attention.attention_utils import (
     build_mask_cache,
     build_mask_slice,
     sample_token_positions,
     ENTRIES_PER_GB,
 )
+from keys_values.attention.base import (
+    DefaultKeysAndValues,
+    MultiHeadSelfAttention,
+    do_softcapping,
+    eager_scaled_dot_product_attention,
+)
+from keys_values.attention.flex_attention import FlexAttentionArgs
 from keys_values.config import Config
-from keys_values.flex_attention import FlexAttentionArgs
 from keys_values.head_model import CrossEntropyOnLogits
 from keys_values.head_model_factory import HeadModelFactory
 from keys_values.kvcache.base import KVCache, KVCacheParams, DefaultKVCache
